Refactor 2/2: fragmented TrackingManager into three classes according to responsibilities
